### 2020-05-22 #2223 @whedon
[PRE REVIEW]: arcos (R) and arcospy (Python): twin software packages to access the DEA ARCOS database from 2006 to 2014
**Submitting author:** @jeffcsauer (<a href="http://orcid.org/0000-0002-8581-6919">Jeffery Sauer</a>)
**Repository:** <a href="https://github.com/jeffcsauer/arcos_arcospy_information" target ="_blank">https://github.com/jeffcsauer/arcos_arcospy_information</a>
**Version:** arcos (1.1); arcospy (1.1.0)
