### 2020-05-22 #2223 @jeffcsauer
Sure, no problem! Thanks for assigning and inviting an editor.

[Link to `arcos` (r) repository](https://github.com/wpinvestigative/arcos)

[Link to `arcospy` (python) repository](https://github.com/jeffcsauer/arcospy)

Both links are also present in the article under the Availability section.
