### 2020-05-29 #2223 @arfon
:wave: @jeffcsauer - @lorenanicole is a new editor and is already handling two JOSS submissions and so won't be able to take this one on. I'm going to mark this as `waitlisted` for now due to the large volume of submissions that have been received by JOSS in the last week but expect we'll be able to identify a handling editor relatively soon.

Thanks in advance for your patience!
