### 2020-06-05 #2223 @bmcfee
@jeffcsauer :wave: Things look generally good so far, so I'll get to finding some reviewers over the next few days.

In the meantime, I've noticed that the [`LICENSE`](https://github.com/wpinvestigative/arcos/blob/master/LICENSE) file in one of the repos doesn't quite line up with the [license listed in the submission repository](https://github.com/jeffcsauer/arcos_arcospy_information/blob/master/LICENSE).  These should be made complete and consistent to satisfy the submission requirements.
