### 2020-06-06 #2223 @jeffcsauer
@bmcfee spoke with co-author and updated LICENSE file in [arcos_arcospy_information](https://github.com/jeffcsauer/arcos_arcospy_information/blob/master/LICENSE) and [arcospy](https://github.com/jeffcsauer/arcospy) to follow the lead of arcos. All LICENSE files across the software repositories and JOSS paper repo page are now consistent.
