### 2020-06-08 #2223 @bmcfee
Thanks @jeffcsauer, however it now appears that these packages are not using a standard OSS license?  The README says "MIT", but the full text of the license should be included in the repository/ies (as per [submission guidelines](https://joss.readthedocs.io/en/latest/submitting.html)).
