### 2020-06-08 #2223 @jeffcsauer
@bmcfee thanks for pointing this out! I spoke again with the co-author and confirmed the MIT + file LICENSE is the appropriate template. I updated the [arcos_arcospy_information ](https://github.com/jeffcsauer/arcos_arcospy_information/blob/master/LICENSE) LICENSE file to avoid confusion. Could you confirm this is now conforms to submission guidelines?
