### 2020-07-08 #2450 @whedon
[REVIEW]: arcos (R) and arcospy (Python): twin software packages to access the DEA ARCOS database from 2006 to 2014
**Submitting author:** @jeffcsauer (<a href="http://orcid.org/0000-0002-8581-6919">Jeffery Sauer</a>)
**Repository:** <a href="https://github.com/jeffcsauer/arcos_arcospy_information" target ="_blank">https://github.com/jeffcsauer/arcos_arcospy_information</a>
**Version:** <v1.1.1>

Comment: No automated tests, but steps to test functionality are clear
