### 2020-07-08 #2450 @jaeyk
@jeffcsauer I will review the package starting today. Could you reorganize `paper.md` in the package git repository so that it could be more consistent with the standard format used in JOSS? An example is [here](https://github.com/ComtekAdvancedStructures/cmstatr/blob/master/paper.md). Specifically, I would like to see a separate section on the statement of need.
