### 2020-07-08 #2450 @jaeyk
@jeffcsauer Also, it would be great, if you could add a short but detailed documentation on the functions provided by the package. An example is [here](https://keyatm.github.io/keyATM/articles/pkgdown_files/Options.html). The R and Python examples provided in the usage section are not user-friendly. You already have the vignettes. Thus, I believe that this is easily doable.
