### 2020-07-09 #2450 @jaeyk
@jeffcsauer Thanks. I plan to review software paper -> documentation -> functionality. The package is already in good shape. My focus will be improving user-interface.
