### 2020-07-10 8e35115
Implementing improvements to paper suggested by jaeyk for JOSS paper
