+++ paper.bib
@@ -26,17 +26,34 @@
 }
 
 @article{diez_2019, 
-  title="The Post Released the DEA’s Data on Pain Pills. Here’s What Local Journalists are Using It For.", 
-  url="https://www.washingtonpost.com/national/2019/08/12/post-released-deas-data-pain-pills-heres-what-local-journalists-are-using-it/&usg=AOvVaw3TrlkTVOxr_jFBdiKIU61G",
-  journal="The Washington Post", 
-  author="Diez, Maria", 
-  year="2019", 
-  month="Aug"} 
+  title = "The Post Released the DEA’s Data on Pain Pills. Here’s What Local Journalists are Using It For.", 
+  url= "https://www.washingtonpost.com/national/2019/08/12/post-released-deas-data-pain-pills-heres-what-local-journalists-are-using-it/&usg=AOvVaw3TrlkTVOxr_jFBdiKIU61G",
+  journal = "The Washington Post", 
+  author = "Diez, Maria", 
+  year = "2019", 
+  month = "Aug"} 
   
 @article{top_2020, 
-  title="Suburbs saw large influx of prescription opioids", 
-  url="https://baltimorefishbowl.com/stories/analysis-marylands-suburbs-saw-a-large-influx-of-prescription-opioids-from-2006-2012/",
-  journal="Bowie Blade-News", 
-  author="Top, Emily, 
-  year="2020", 
-  month="Jan"} 
+  title = "Suburbs saw large influx of prescription opioids", 
+  url = "https://baltimorefishbowl.com/stories/analysis-marylands-suburbs-saw-a-large-influx-of-prescription-opioids-from-2006-2012/",
+  journal = "Bowie Blade-News", 
+  author = "Top, Emily, 
+  year = "2020", 
+  month = "Jan"} 
+  
+@article{reisman_2009,
+	title = "Prescription opioid usage and abuse relationships: An evaluation of state prescription drug monitoring program efficacy",
+	author = "Reisman,  M and Shenoy, P and Atherly, A and Flowers, C",
+	url = "https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3865068/",
+	journal = "Substance Abuse",
+	doi = "10.4137/sart.s2345",
+	year = "2009"}
+	
+@article{Kenan2012,
+	title = "Trends in prescriptions for oxycodone and other commonly used opioids in the United States, 2000-2010",
+	author = "Kenan, K and Mack, K and Paulozzi, L",
+	url = "https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3659213/",
+	journal = "Open Medicine",
+	year = "2012"
+}
+
