+++ paper.md
@@ -25,34 +25,80 @@ affiliations:
    index: 1
  - name: University of Maryland, College Park, Department of Geographical Sciences
    index: 2
-date: 01 March 2020
+date: 09 July 2020
 bibliography: paper.bib
 ---
 
 # Summary
 
-The ongoing Opioid Crisis in the United States presents a complex sociomedical epidemic. Researchers, journalists, and government agencies are actively investigating the myriad impacts of the crisis. One powerful tool for understanding trends in prescription opioid distribution is the Drug Enforcement Agency's (DEA) Automation of Reports and Consolidated Orders System (ARCOS). ARCOS tracks the commercial distribution of controlled substances in the United States. The data is highly detailed, tracking commercial origin, pharmacy order frequency, point-of-sale distribution, and more. For a variety of reasons ranging from patient confidentiality to protecting trade secrets, access to the ARCOS data is normally restricted. Recent litigation efforts by *The Washington Post*, HD Media, and local journalists allowed for the public release of a large portion of the ARCOS database from 2006 to 2012. `arcos` and `arcospy` are open source API wrappers in R and Python, respectively, that allow researchers and interested citizens to easily access the ARCOS database.
+The ongoing Opioid Overdose Crisis in the United States presents a complex sociomedical epidemic. Researchers, journalists, and government agencies are actively investigating the myriad impacts of the crisis. One powerful tool for understanding trends in prescription opioid distribution is the Drug Enforcement Agency's (DEA) Automation of Reports and Consolidated Orders System (ARCOS). ARCOS tracks the commercial distribution of controlled substances in the United States. The data is highly detailed, tracking commercial origin, pharmacy order frequency, point-of-sale distribution, and more. For a variety of reasons ranging from patient confidentiality to protecting trade secrets, access to sub-state ARCOS data is normally restricted. Recent litigation efforts by *The Washington Post*, HD Media, and local journalists allowed for the public release of a large portion of the ARCOS database from 2006 to 2012. `arcos` and `arcospy` are open source API wrappers in `R` and `Python`, respectively, that allow researchers and interested citizens to easily access this newly available portion of the ARCOS database.
 
-## Data availability
+## Statement of need
 
-In raw format, the available portion of the ARCOS database is more than 130 gigabytes and includes several hundred columns. Thus, `arcos` and `arcospy` are meant to simplify access to the database and gather relevant measures of prescription opioid distribution. Data can be gathered at the pharmacy, distributor, county, or state as the geographic unit of analysis. Depending on the geographic level, there may be raw, summarized, or supplemental data available. For example, the `county_raw()` command returns each individual ARCOS record for a given county from 2006 to 2012. However, the `summarized_county_annual()` command returns the annual summarized totals for a given county for each year of 2006 to 2012. Full documentation on how data is collected by the DEA is available in the [ARCOS Registrant Handbook](https://www.deadiversion.usdoj.gov/arcos/handbook/full.pdf). Supplemental commands return relevant data - such as county population - gathered from the American Community Survey.
[…]
