+++ README.md
@@ -26,7 +26,42 @@ Build status is evaluated via CRAN for `arcos` and travis for `arcospy`.
 
 ## Features
 
-**arcos** and **arcospy** offer 24 functions for users to access DEA ARCOS data using pharmacies, distributors, counties, or states as the unit of analysis. These functions have the exact same name in both pieces of software to rapidly switch between languages if the need arises (i.e. if a certain type of analysis is available in R or Python).
+**arcos** and **arcospy** offers nearly 30 functions for users to access DEA ARCOS data using pharmacies, distributors, counties, or states as the unit of analysis, as well as useful supplementary information. These functions have the exact same name in both `R` and `Python`, allowing users to rapidly switch between languages if the need arises (i.e. if a certain type of analysis is available only in R or Python).
+
+
+__Functions and the available datasets (Read the [reference page](https://wpinvestigative.github.io/arcos/reference/index.html) for more info):__
+
+| Function                                                                  | What                                                                            | Type         | Years       | Drugs                   | Buyers                                             |
+|---------------------------------------------------------------------------|---------------------------------------------------------------------------------|--------------|-------------|-------------------------|----------------------------------------------------|
+| [buyer_addresses()](https://wpinvestigative.github.io/arcos/reference/buyer_addresses.html)                       | Get DEA designated addresses for each pharmacy                                  | Raw          |             | All                     | All                                                |
+| [buyer_details()](https://wpinvestigative.github.io/arcos/reference/buyer_details.html)                           | Get monthly summarized pill totals by county                                    | Summarized   | 2006 - 2014 | Oxycodone & Hydrocodone | Retail Pharmacy, Chain Pharmacy, and Practitioners |
+| [buyer_list()](https://wpinvestigative.github.io/arcos/reference/buyer_list.html)                                 | Get list of business types listed in the BUYER_BUS_ACT in the ARCOS database    | Raw          | 2006 - 2014 |                         | All                                                |
+| [combined_buyer_annual()](https://wpinvestigative.github.io/arcos/reference/combined_buyer_annual.html)           | Get annual total pills for each buyer (pharmacy, etc) in a county               | Summarized   | 2006 - 2014 | Oxycodone & Hydrocodone | Retail Pharmacy, Chain Pharmacy, and Practitioners |
+| [combined_buyer_monthly()](https://wpinvestigative.github.io/arcos/reference/combined_buyer_monthly.html)         | Get annual total pills for each buyer (pharmacy, etc) in a county               | Summarized   | 2006 - 2014 | Oxycodone & Hydrocodone | Retail Pharmacy, Chain Pharmacy, and Practitioners |
+| [county_list()](https://wpinvestigative.github.io/arcos/reference/county_list.html)                               | Get list of counties and states and fips codes represented in the ARCOS data    | Raw          | 2006 - 2014 |                         |                                                    |
+| [county_population()](https://wpinvestigative.github.io/arcos/reference/county_population.html)                   | Get annual population for counties between 2006 and 2012                        | Supplemental | 2006 - 2014 | Oxycodone & Hydrocodone | Retail Pharmacy, Chain Pharmacy, and Practitioners |
+| [county_raw()](https://wpinvestigative.github.io/arcos/reference/county_raw.html)                                 | Download raw prescription data for specified county (by state and county names) | Raw          |             | Oxycodone & Hydrocodone | Retail Pharmacy, Chain Pharmacy, and Practitioners |
[…]
