+++ paper.bib
@@ -38,7 +38,7 @@
   title = "Suburbs saw large influx of prescription opioids", 
   url = "https://baltimorefishbowl.com/stories/analysis-marylands-suburbs-saw-a-large-influx-of-prescription-opioids-from-2006-2012/",
   journal = "Bowie Blade-News", 
-  author = "Top, Emily, 
+  author = "Top, Emily", 
   year = "2020", 
   month = "Jan"
 }
