### 2020-07-10 3348c14
Rewriting a couple sentences
