### 2020-07-10 063eae4
Fixing table formatting in paper.md
