### 2020-07-10 30ba3aa
Removing excess doi in paper.bib
