+++ paper.bib
@@ -3,7 +3,7 @@
   author = {Stopka, TJ and Amaravadi, H and Kaplan, AR and Hoh, R and Bernson, D and Chui, KKH and Land, T and Walley, AY and LaRochelle, MR and Rose, AJ},
   journal = {International Journal of Drug Policy},
   year = 2019,
-  doi = {doi:10.1016/j.drugpo.2019.03.024},
+  doi = {10.1016/j.drugpo.2019.03.024},
   url = {https://www.sciencedirect.com/science/article/pii/S0955395919300908}
 }
 
@@ -12,7 +12,7 @@
   author = {Gilson, A and Ryan, K and Joranson, D and Dahl, J},
   journal = {Journal of Pain and Symptom Management},
   year = 2004,
-  doi = {doi:10.1016/j.jpainsymman.2004.01.003},
+  doi = {10.1016/j.jpainsymman.2004.01.003},
   url = {https://www.jpsmjournal.com/article/S0885-3924(04)00182-4/fulltext}
 }
 
@@ -21,7 +21,7 @@
   author = {Joranson, D and Ryan, K and Gilson, A and Dahl, J},
   journal = {Journal of the American Medical Association},
   year = 2000,
-  doi = {doi:10.1001/jama.283.13.1710},
+  doi = {10.1001/jama.283.13.1710},
   url = {https://jamanetwork.com/journals/jama/fullarticle/vol/283/pg/1710}
 }
 
