### 2020-07-10 9130f34
Fixing visual glitch with bullet points
