### 2020-07-10 #2450 @jeffcsauer
@jaeyk wanted to ping you as I have incorporated your suggestions into the paper. A summary of changes include:

- Addition of a 'Statement of Need' section. This contextualizes the dataset as a major addition to analyzing prescription opioid data due to the dearth of existing datasets on the topic. The section also states the core goals of the software (i.e. data access, relevant metrics, and language flexibility).
- Expansion of 'Data Availability' section to include some examples of 'Basic Usage'.
- Slight re-organization of paper structure to improve readability (i.e. 'API Structure' section now comes before 'Data Availability and Basic Usage' section).
- Addition of link to shared repository under 'Availability' section.
- A few grammar and writing changes to improve readability.

Regarding your concern about users finding the functions, I have added a table of the available functions on the [shared arcos/arcospy repository](https://github.com/jeffcsauer/arcos_arcospy_information) as well as linking to the table/repository in the paper (appears in a 'Data Availability and Basic Usage' and 'Availability' sections). Let me know if this is helpful or if you were thinking of a different solution!
