### 2020-07-10 #2450 @jaeyk
@jeffcsauer Thanks for the quick update. They look great. I have a couple more minor suggestions:

- Provide references at the bottom of the `paper.md`. (You already have in-text citations and have a section heading for citations. So, I believe that you plan to do this.)
- Provide more structure to contributions. Take a look at [this example](https://github.com/keyATM/keyATM/issues). For instance, when create issues, are issues about bugs/features? When users report bugs/features, what kind of the format they should follow? Are they required to report their environment (e.g., session information)? Should they provide a minimal reproducible example? Having this sort of clear guideline helps users to create issues also helps you to include issues in a development roadmap.
