### 2020-07-13 66906fc
Adding bug and feature request templates
