+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,1 +1,20 @@
-If you have an idea for `arcos` or `arcospy`, please describe your idea here! The maintaters of `arcos` and `arcospy` are actively interested in expanding the capabilities of the software, improving user experience, and filling in important documentation. 
+---
+name: Feature request
+about: Suggest an idea for this project
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+**Is your feature request related to a problem? Please describe.**
+A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
+
+**Describe the solution you'd like**
+A clear and concise description of what you want to happen.
+
+**Describe alternatives you've considered**
+A clear and concise description of any alternative solutions or features you've considered.
+
+**Additional context**
+Add any other context or screenshots about the feature request here.
