+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -11,11 +11,8 @@ assignees: ''
 A clear and concise description of what the bug is.
 
 **To Reproduce**
-Steps to reproduce the behavior:
-1. Go to '...'
-2. Click on '....'
-3. Scroll down to '....'
-4. See error
+
+Please paste a reproducible example of the bug you are encountering.
 
 **Expected behavior**
 A clear and concise description of what you expected to happen.
@@ -24,15 +21,20 @@ A clear and concise description of what you expected to happen.
 If applicable, add screenshots to help explain your problem.
 
 **Desktop (please complete the following information):**
- - OS: [e.g. iOS]
- - Browser [e.g. chrome, safari]
- - Version [e.g. 22]
 
-**Smartphone (please complete the following information):**
- - Device: [e.g. iPhone6]
- - OS: [e.g. iOS8.1]
- - Browser [e.g. stock browser, safari]
- - Version [e.g. 22]
+- Software: are you using R or Python?
 
-**Additional context**
-Add any other context about the problem here.
+- Python information:
+```python
+>>> import os; print(os.name, os.sys.platform)
+>>> import sys; print(sys.version)
+>>> import arcospy; print(arcospy.__version__)
+```
+
+- R information:
+```r
+sessionInfo()
+packageVersion("arcos")
+```
+
+* What arcos/arcospy commands are you using?
