### 2020-07-14 #2450 @jeffcsauer
@jaeyk thanks for the additional input! I will am working on bullet point 2 and will get back to you in a few days when done.

Quick question regarding bullet point 1. Should I add an additional 'Citations' section in the `paper.md` file that is commented out? The above article preview generated by Whedon appears to be generating a correct and completed 'Citations' section. Apologies for my confusion on this point.
