### 2020-07-15 052e008
Updating contributing guidelines as well as issue and feature templates
