+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -7,14 +7,28 @@ assignees: ''
 
 ---
 
-**Is your feature request related to a problem? Please describe.**
-A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
+Thank you for taking the time to suggest a new feature for `arcos` and `arcospy`! Please see the guidelines outlined in the contributing.md document before filling out the following form.
+
+**Is this a feature request related to the underlying ARCOS data or arcos/arcospy API?**
+Please identify if your feature request seeks something in the ARCOS data or is a new function for the arcos/arcospy API.
+
+**Would this feature help you obtain a measurement or data product that has precedent in the appropriate literature?**
+If so, please link or identify the relevant literature. If not, please briefly explain why this measurement or data product would be useful to a wider audience.
+
+**Is your feature request related to an existing problem? Please describe.**
+A clear and concise description of what the problem is. 
+Ex. "The current function does X, which is a problem because..."
 
 **Describe the solution you'd like**
-A clear and concise description of what you want to happen.
+A clear and concise description of what you want to happen with the feature. This may inlcude:
+- New field(s)
+- Resulting data type(s)
+- Geographic level(s) of the data (if applicable)
+- Timespan of the data (if applicable)
 
 **Describe alternatives you've considered**
 A clear and concise description of any alternative solutions or features you've considered.
+Ex. Have you tried creating the function locally using existing arcos/arcospy commands?
 
 **Additional context**
 Add any other context or screenshots about the feature request here.
