+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -7,11 +7,29 @@ assignees: ''
 
 ---
 
+Thank you for taking the time to report a bug for the `arcos` and `arcospy` software! Please fill out as much of the following template as possible. Feel free to delete excess text after you have provided the relevant information.
+
+**Software information**
+
+- Software: are you using R or Python?
+
+- If using Python:
+```python
+>>> import os; print(os.name, os.sys.platform)
+>>> import sys; print(sys.version)
+>>> import arcospy; print(arcospy.__version__)
+```
+
+- If using R:
+```r
+sessionInfo()
+packageVersion("arcos")
+```
+
 **Describe the bug**
 A clear and concise description of what the bug is.
 
-**To Reproduce**
-
+**Code to Reproduce**
 Please paste a reproducible example of the bug you are encountering.
 
 **Expected behavior**
@@ -20,21 +38,3 @@ A clear and concise description of what you expected to happen.
 **Screenshots**
 If applicable, add screenshots to help explain your problem.
 
-**Desktop (please complete the following information):**
-
-- Software: are you using R or Python?
-
-- Python information:
-```python
->>> import os; print(os.name, os.sys.platform)
->>> import sys; print(sys.version)
->>> import arcospy; print(arcospy.__version__)
-```
-
-- R information:
-```r
-sessionInfo()
-packageVersion("arcos")
-```
-
-* What arcos/arcospy commands are you using?
