### 2020-07-15 #2450 @jeffcsauer
@jaeyk No need to apologize! Thanks for the quick response.

Following up on your suggestion to "Provide more structure to the `contributing.md`...", I have carried out the following changes:

- Updated `contributing.md` to describe the purpose of the `arcos_arcospy_information` repository, namely centralizing and standardizing issues/feature requests. I also added guidelines to help identify whether an issue/feature request applies to the core ARCOS data or the `arcos` and `arcospy` software. Added detail to several sub-sections of `contributing.md` in general.
- Adopted Github templates on the shared `arcos_arcospy_information` repository. As now explained in `contributing.md`, issues/feature requests posted to the separate `arcos` and `arcospy` repositories will now be cross-posted and standardized on the shared `arcos_arcospy_information` repository.
- Notifications are set up to notify the maintainers of the R/Python software when an issue is raised on the opposite repository. I have taken on the responsibility to cross-post the issues to the shared `arcos_arcospy_information` repository.
