### 2020-07-31 #2450 @jaeyk
@jeffcsauer Could you rename [the Articles section in the tab in the reference page](https://wpinvestigative.github.io/arcos/articles/per-capita-pharmacies.html) for something more intuitive (e.g. User's Guide)? As of now, it's not very clear what it means.
