### 2020-07-31 #2450 @jaeyk
@jeffcsauer  It would be great, if you include some unit tests. For instance,

```r
states <- combined_buyer_annual(key="WaPo")

```
