### 2020-07-31 #2450 @jaeyk
Functionality checks for the R version

- [ ] combined_buyer_annual()
- [ ] total_pharmacies_county()
- [ ] drug_list()
- [ ] drug_county_biz()
- [ ] total_pharmacies_state()
- [ ] county_population()
- [ ] buyer_address()
- [ ] pharmacy_raw()
