### 2020-07-31 #2450 @jaeyk
@jeffcsauer Could you provide the vignette for the Python version package? Currently, the reference in the Python package is linked to the vignette section of the R version package.
