### 2020-08-01 #2450 @jeffcsauer
@jaeyk thanks for the additional comments! A few follow-up points/questions:
- I have opened [an issue](https://github.com/wpinvestigative/arcos/issues/7) on the R/arcos repository to field possible renames for the button. I agree, this button should likely be renamed to something more intuitive.
- I have re-arranged (and added) [additional notebooks to the R/arcospy repository](https://github.com/jeffcsauer/arcospy/tree/master/docs) (more to come!). When you say 'vignette for the Python version', what kind of vignette were you envisioning (e.g. read the docs, notebooks, or something else)? The primary reason for the many links back to the R/arcos docs is that R/arcos was created first with the python version coming later. Early discussions between the R and python authors wanted to avoid duplication of documentation given the primacy and quality of the R/arcos materials. I would be happy to include a prominent statement explaining this to users.
