### 2020-08-02 #2450 @jaeyk
@jeffcsauer

re1: Great!

re2: I appreciate the detailed R documentation but it would be great if there's R & Python side-by-side vignette as that would help users to see how parallel the two packages are. For instance, users might have questions about whether the Python version has identical function names, arguments, etc. So, provide a minimum vignette that clarifies this potential confusion.

See the vignette page of the [{prophet}](https://facebook.github.io/prophet/docs/saturating_forecasts.html#forecasting-growth) package which is also available both in R and Python. The usage of the two packages are almost identical but you can also notice that a slight difference. For instance, `prophet()` in R is equivalent to `Prophet()` in Python.
