### 2020-08-02 #2450 @jeffcsauer
@jaeyk

Ah thank you so much for the clarification and example! I was misunderstanding the original point. Will do - I will work on a brief vignette that does side-by-side comparison and will make it prominently displayed on the shared arcos/arcospy repository, as well as linking on the specific software. I will ping you when it is done!
