### 2020-08-07 b36744c
Adding dual software vignette and updating landing page
