+++ docs/R_Python_arcos_vignette.Rmd
@@ -1,0 +1,267 @@
+---
+title: "Accessing ARCOS data in R and Python"
+author: "Authors: arcos and arcospy developers"
+date: "Last updated: August 7, 2020"
+output:
+  html_document:
+    code_folding: show
+    theme: readable
+    toc: yes
+  pdf_document: default
+---
+
+```{r setup, include=FALSE}
+knitr::opts_chunk$set(echo = TRUE)
+library(reticulate)
+use_python("C:/Users/USER/AppData/Local/Programs/Python/Python38/python.exe")
+```
+
+## Installation
+
+The original `arcos` software was written in R by members of the data reporting team (Steven Rich, Andrew Ba Tran, Aaron Williams, Jason Holt) at *The Washington Post*. Interested researchers at the University of Maryland (Jeffery Sauer, Dr. Taylor M. Oshan) translated the software to python and called it `arcospy`. Both software offer the same functionality as an API wrapper to query the ARCOS data (hosted [here](https://arcos-api.ext.nile.works/__swagger__/)). 
+
+### R
+
+The latest release of `arcos` is available via CRAN, while the development version is available from GitHub.
+
+```{r, eval=FALSE, warning=FALSE}
+
+# R
+
+# CRAN
+
+install.packages("arcos")
+
+# Github
+
+# install.packages("devtools")
+devtools::install_github('wpinvestigative/arcos')
+
+```
+
+The R dependencies are `string`, `stringr`, `magrittr`, `jsonlite`, `dplyr`, `urltools`, and `vroom`.
+
+### Python
+
+Python is available on PyPI and can be installed via `pip`.
+
+```{python, eval = FALSE, warning = FALSE}
+
+# Python
+
+pip install arcospy
+
+```
+
+The Python dependencies are `pandas` and `requests`. 
+
+## Quick start
+
+All commands share the name name between `arcos` and `arcospy`. Users can choose between commands that return raw data, commands that return summarized data, and commands that return potentially relevant supplemental data. Note that commands that return raw data may take significantly longer to run due to large size of the underlying datasets.
+
+Once users have chosen a command that suits their needs, the general format of the API calls for the geographic areas of interest (such as county and state) as well as an API key. The default key is "WaPo" (additional keys available [here](https://github.com/wpinvestigative/arcos-api/blob/master/keys/keys.txt)). Outputs from all of the functions are delivered in popular formats, specifically `data.frames` in R and `pandas.DataFrame` in Python. We now demonstrate generic functionality with the `state_population` command to get annual population values for states between 2006 and 2012. 
+
+### R API
+
+```{r, warning = FALSE}
+
+# R
+
+library(arcos)
+statepop <- state_population(state = "WV", key = "WaPo")
+head(statepop)
+
+```
+
+Help for the R `arcos` package can be retrieved via `?arcos`. Additional documentation on the R API is available via [CRAN](https://cran.r-project.org/web/packages/arcos/arcos.pdf).
+
+### Python API
+
+```{python, warning = FALSE}
+
+# Python
+
+import arcospy
+statepop = arcospy.state_population(state = "WV", key = "WaPo")
+statepop.head()
+
+```
+
+Help for the Python `arcospy` package can be retrieved via `help(arcospy)`. Additional documentation on the Python API is available via [Github](https://github.com/jeffcsauer/arcospy).
+
+## Example use: county-level analysis over time 
+
+### Gathering data
+
[…]
