### 2020-08-07 9062ef6
Updating .pdf and .html render
