### 2020-08-07 #2450 @jeffcsauer
@jaeyk

I have [created a vignette](https://github.com/jeffcsauer/arcos_arcospy_information/tree/master/docs) that does provides side-by-side comparison for installing the package, a quick start example, and a more involved example use case. I also highlight key considerations for users of the data (as highlighted in the paper) and provide some additional useful links. The vignette resides on the shared arcos/arcospy repository and is prominently placed in the first text block. Thanks again for this suggestion - it is indeed an easier starting point than going back-and-forth! Let me know what else I can do!
