### 2020-09-05 #2450 @edonnachie
Before I get to the package itself, here are a few initial comments on the paper:

__Summary:__

- It would be helpful to explain what the Opoid Overdose Crisis is, perhaps with link to the [CDC](https://www.cdc.gov/opioids/strategy.html). It's not clear to me whther the overdoses are accidental (due to overuse of this drug therapy) or suicide-related (due to the controls being too lax).

- You write that data was released until 2012. You should mention here that the data for 2013-2014 is now also available.

- You write that patient confidentiality is one reason that access to ARCOS is usually restricted. If this were true, then providing wide access to this data would be legally and ethically problematic (at least from a european perspective). However, I would tend to the view that the risk of revealing information about patients is negligible. Either way, this needs some explaining.

- Both from a technical and a data protection perspective, it would seem important to explain what is behind the API: How and where are the data stored? Who controls and maintains this database? Are reqiests logged, and if so, are personally identifiable data kept (e.g. IP address)?

__Statement of Need:__

- You should explain why the ARCOS database is useful. How does this compare with other sources of prescribing data (pharmacy data centres, medicaid etc)? I suspect that the main advantage lies in the analysis of how market dynamics facilitated the epidemic.

- Repeated word: "In raw format, the the ARCOS database "

__Data Availability and Basic Usage:__

- Should "from 2006 to 2012" not be "from 2006 to 2014"?

__Conclusion:__

- The statement "while still maintaining individual medical privacy " is confusing, see above.

- What you don't mention is that the data provide a different perspective to the medical claims data that would usually be used for such analyses. If I understand correctly, you're not looking at who received an opoid, but at how these substances travelled across the supply chain (and in what quantity). I think this is an important and interesting distinction to be made.
