### 2020-09-05 #2450 @jeffcsauer
Thanks for these detailed initial comments! I will incorporate them over the next couple of days and tag you when done.
