### 2020-09-07 bdd6008
Responding to comments by edonnachie
