+++ paper.bib
@@ -1,3 +1,58 @@
+@article{rudd,
+  title = {Opioid overdose deaths and potentially inappropriate opioid prescribing practices (pip): A spatial epidemiological study},
+  author = {Stopka, TJ and Amaravadi, H and Kaplan, AR and Hoh, R and Bernson, D and Chui, KKH and Land, T and Walley, AY and LaRochelle, MR and Rose, AJ},
+  journal = {International Journal of Drug Policy},
+  year = 2019,
+  doi = {10.1016/j.drugpo.2019.03.024},
+  url = {https://www.sciencedirect.com/science/article/pii/S0955395919300908}
+}
+
+@article{cdcweb, 
+  title = "Prescribing Practices", 
+  url= "https://www.cdc.gov/drugoverdose/data/prescribing/prescribing-practices.html",
+  journal = "Centers for Disease Control and Prevention, National Center for Injury Prevention and Control", 
+  year = "2019", 
+  month = "Aug"
+}
+
+@article{arcosprivacy, 
+  title = "Privacy Impact Assessment for the Registrant Information Consolidated System", 
+  author = "Grubbs, Preston"
+  url= "https://www.dea.gov/sites/default/files/2018-06/rics_%20pia_060414.pdf",
+  journal = "Drug Enforcement Administration", 
+  year = "2014", 
+  month = "June"
+}
+
+@article{Guy2017,
+  title = {Vital Signs: Changes in Opioid Prescribing in the United States, 2006–2015}},
+  author = {Guy, Gery P. and Zhang, Kun and Bohm, Michele K. and Losby, Jan and Lewis, Brian and Young, Randall and Murphy, Louise B. and Dowell, Deborah},
+  journal = {Morbidity and Mortality Weekly Report},
+  month = July,
+  year = {2017}
+  doi = {10.15585/mmwr.mm6626a4},
+  url = {http://www.cdc.gov/mmwr/volumes/66/wr/mm6626a4.htm},
+}
+
+@article{VanZee2009,
+  author = {Van Zee, Art},
+  title = {The promotion and marketing of oxycontin: commercial triumph, public health tragedy},
+  journal = {American journal of public health},
+  year = {2009}
+  doi = {10.2105/AJPH.2007.131714},
+  url = {http://www.ncbi.nlm.nih.gov/pubmed/18799767 http://www.pubmedcentral.nih.gov/articlerender.fcgi?artid=PMC2622774},
+}
+
+@article{ko2020,
+  title = {Vital Signs: Prescription Opioid Pain Reliever Use During Pregnancy — 34 U.S. Jurisdictions, 2019},
+  author = {Ko, Jean Y. and D'Angelo, Denise V. and Haight, Sarah C. and Morrow, Brian and Cox, Shanna and Salvesen von Essen, Beatriz and Strahan, Andrea E. and Harrison, Leslie and Tevendale, Heather D. and Warner, Lee and Kroelinger, Charlan D. and Barfield, Wanda D.},
+  journal = {Morbidity and Mortality Weekly Report},
+  month = July,
+  year = 2020,
+  doi = {10.15585/mmwr.mm6928a1},
+  url = {http://www.cdc.gov/mmwr/volumes/69/wr/mm6928a1.htm?s{\_}cid=mm6928a1{\_}w},
+}
+
 @article{stopka,
   title = {Opioid overdose deaths and potentially inappropriate opioid prescribing practices (pip): A spatial epidemiological study},
   author = {Stopka, TJ and Amaravadi, H and Kaplan, AR and Hoh, R and Bernson, D and Chui, KKH and Land, T and Walley, AY and LaRochelle, MR and Rose, AJ},
