+++ paper.md
@@ -31,13 +31,15 @@ bibliography: paper.bib
 
 # Summary
 
-The ongoing Opioid Overdose Crisis in the United States presents a complex sociomedical epidemic. Researchers, journalists, and government agencies are actively investigating the myriad impacts of the crisis. One powerful tool for understanding trends in prescription opioid distribution is the Drug Enforcement Agency's (DEA) Automation of Reports and Consolidated Orders System (ARCOS). ARCOS tracks the commercial distribution of controlled substances in the United States. The data is highly detailed, tracking commercial origin, pharmacy order frequency, point-of-sale distribution, and more. For a variety of reasons ranging from patient confidentiality to protecting trade secrets, access to sub-state ARCOS data is normally restricted. Recent litigation efforts by *The Washington Post*, HD Media, and local journalists allowed for the public release of a large portion of the ARCOS database from 2006 to 2012. `arcos` and `arcospy` are open source API wrappers in `R` and `Python`, respectively, that allow researchers and interested citizens to easily access this newly available portion of the ARCOS database.
+In the early 2000s governmental agencies across the United States began to observe increases in the number of all-cause opioid-involved deaths [@rudd]. The Centers for Disease Control (CDC) describe this Opioid Overdose Epidemic as occuring in three waves, with the first wave attributed to the widespread distribution of prescription opioids, reaching an opioid prescribing rate as high as 81.3 per 100 persons in 2012 [@cdcweb]. While waves two and waves three of the Opioid Overdose Epidemic are largely attributed to illicit opioids like heroin and fentanyl, prescribing rates and prescription misuse remain at a high level [@Guy2017;@ko2020]. 
+
+Researchers, journalists, and government agencies are still actively investigating the myriad impacts of prescription opioids on the Opioid Overdose Epidemic. One powerful tool for understanding prescription opioid distribution is the Drug Enforcement Administration's (DEA) Automation of Reports and Consolidated Orders System (ARCOS). ARCOS tracks the commercial distribution of controlled substances in the United States. The data is highly detailed, tracking commercial origin, pharmacy order frequency, point-of-sale distribution, and more. For a variety of reasons ranging from patient confidentiality to protecting trade secrets, access to sub-state ARCOS data is available only for approved requests (e.g. research or litigation) [@arcosprivacy]. Recent litigation efforts by *The Washington Post*, HD Media, and local journalists allowed for the public release of an anonymized, large portion of the ARCOS database from 2006 to 2012, with additional data for 2013 and 2014 now also available. `arcos` and `arcospy` are open source API wrappers in `R` and `Python`, respectively, that allow researchers and interested citizens to easily access this newly available portion of the ARCOS database.
 
 ## Statement of Need
 
-Previously, researchers wanting to use ARCOS data relied on what was made available by the DEA, typically in the form of state-level estimates, or submitted special access data requests to the DEA [@Kenan2012;@reisman_2009]. While alternative data on prescription records are offered by the Centers for Medicare & Medicaid Services in the [Medicare Provider Utilization and Payment Datasets](https://www.cms.gov/Research-Statistics-Data-and-Systems/Statistics-Trends-and-Reports/Medicare-Provider-Charge-Data), this data pertains to a specific sample of the population and spans a different set of years (2011 to 2017). The release of national, longitudinal, sub-state ARCOS data is a major contribution for researchers interested in the distribution of prescription opioids and the subsequent sociomedical impacts.
[…]
