+++ paper.bib
@@ -18,7 +18,7 @@
 @article{arcosprivacy, 
   title = "Privacy Impact Assessment for the Registrant Information Consolidated System", 
   author = "Grubbs, Preston"
-  url= "https://www.dea.gov/sites/default/files/2018-06/rics_%20pia_060414.pdf",
+  url = "https://www.dea.gov/sites/default/files/2018-06/rics_%20pia_060414.pdf",
   journal = "Drug Enforcement Administration", 
   year = "2014", 
   month = "June"
