+++ paper.bib
@@ -17,15 +17,15 @@
 
 @article{arcosprivacy, 
   title = "Privacy Impact Assessment for the Registrant Information Consolidated System", 
-  author = "Grubbs, Preston"
-  url = "https://www.dea.gov/sites/default/files/2018-06/rics_%20pia_060414.pdf",
+  author = "Grubbs, Preston",
   journal = "Drug Enforcement Administration", 
   year = "2014", 
   month = "June"
+  url = "https://www.dea.gov/sites/default/files/2018-06/rics_ 20pia_060414.pdf",
 }
 
 @article{Guy2017,
-  title = {Vital Signs: Changes in Opioid Prescribing in the United States, 2006–2015}},
+  title = {Vital Signs: Changes in Opioid Prescribing in the United States, 2006–2015},
   author = {Guy, Gery P. and Zhang, Kun and Bohm, Michele K. and Losby, Jan and Lewis, Brian and Young, Randall and Murphy, Louise B. and Dowell, Deborah},
   journal = {Morbidity and Mortality Weekly Report},
   month = July,
