### 2020-09-07 c845a31
Changing URL to landing page of all privacy docs
