### 2020-09-07 40dff66
Updating more citations
