+++ paper.bib
@@ -20,8 +20,8 @@
   author = "Grubbs, Preston",
   journal = "Drug Enforcement Administration", 
   year = "2014", 
-  month = "June"
-  url = "https://www.dea.gov/privacy-impact-assessment-and-management-information-systems",
+  month = "June",
+  url = "https://www.dea.gov/sites/default/files/2018-06/rics_%20pia_060414.pdf"
 }
 
 @article{Guy2017,
@@ -29,18 +29,18 @@
   author = {Guy, Gery P. and Zhang, Kun and Bohm, Michele K. and Losby, Jan and Lewis, Brian and Young, Randall and Murphy, Louise B. and Dowell, Deborah},
   journal = {Morbidity and Mortality Weekly Report},
   month = July,
-  year = {2017}
+  year = {2017},
   doi = {10.15585/mmwr.mm6626a4},
-  url = {http://www.cdc.gov/mmwr/volumes/66/wr/mm6626a4.htm},
+  url = {http://www.cdc.gov/mmwr/volumes/66/wr/mm6626a4.htm}
 }
 
 @article{VanZee2009,
   author = {Van Zee, Art},
   title = {The promotion and marketing of oxycontin: commercial triumph, public health tragedy},
   journal = {American journal of public health},
-  year = {2009}
+  year = {2009},
   doi = {10.2105/AJPH.2007.131714},
-  url = {http://www.ncbi.nlm.nih.gov/pubmed/18799767 http://www.pubmedcentral.nih.gov/articlerender.fcgi?artid=PMC2622774},
+  url = {http://www.ncbi.nlm.nih.gov/pubmed/18799767 http://www.pubmedcentral.nih.gov/articlerender.fcgi?artid=PMC2622774}
 }
 
 @article{ko2020,
@@ -50,7 +50,7 @@
   month = July,
   year = 2020,
   doi = {10.15585/mmwr.mm6928a1},
-  url = {http://www.cdc.gov/mmwr/volumes/69/wr/mm6928a1.htm?s{\_}cid=mm6928a1{\_}w},
+  url = {http://www.cdc.gov/mmwr/volumes/69/wr/mm6928a1.htm?s{\_}cid=mm6928a1{\_}w}
 }
 
 @article{stopka,
