### 2020-09-08 7fba7ab
Updating a few sentences
