+++ paper.md
@@ -47,7 +47,7 @@ In raw format, the ARCOS database is more than 130 gigabytes and includes severa
 
 ## API Structure
 
-The `arcos` and `arcospy` API is [publically available](https://arcos-api.ext.nile.works/__swagger__/) and hosted using the [OpenAPI specification](https://swagger.io/specification/). The primary maintainers of API database are members of the Data Reporting Team at *The Washington Post*. A key is required to use the API. The standard key is `WaPo` and additional keys may be sourced from [the Github repository](https://github.com/wpinvestigative/arcos).
+The `arcos` and `arcospy` API is [publically available](https://arcos-api.ext.nile.works/__swagger__/) and hosted using the [OpenAPI specification](https://swagger.io/specification/). The primary maintainers of API database are members of the Data Reporting Team at *The Washington Post*. A key is required to use the API. The standard key is `WaPo` and additional keys may be sourced from [the Github repository](https://github.com/wpinvestigative/arcos). Guidelines on using the API are available from [*The Washington Post*](https://www.washingtonpost.com/national/2019/07/18/how-download-use-dea-pain-pills-database/).
 
 All commands share the same name between `arcos` and `arcospy`. This allows users to rapidly switch between languages if the need arises. Outputs from all of the functions are delivered in popular formats - `data.frames` in `R` and `pandas.DataFrame` in `python` - to enable statistical, spatial, network, or other types of analysis.
 
@@ -99,7 +99,7 @@ HillSummarized.head()
 
  Given the number of records, users should anticipate that commands querying for raw data will take longer than commands querying for summarized data. Full documentation on how data is collected by the DEA is available in the [ARCOS Registrant Handbook](https://www.deadiversion.usdoj.gov/arcos/handbook/full.pdf). `arcos` and `arcospy` also include supplemental commands that relevant data - such as county population - gathered from the American Community Survey. A description of each of the functions currently offered, as well as examples in `R` and `Python` demonstrating functionality, are available on [the shared `arcos` and `arcospy` repository](https://github.com/jeffcsauer/arcos_arcospy_information).
 
-There are several ways to conceptualize the unit of analysis for opioids from the present data. These include the total number of records, the total number of all opioid pills, the total number of specific opioid pills (i.e. oxycodone versus hydrocodone), or the total amount (in weight) of all or specific opioid pills. Other common units of analysis that may be of interest include morphine milligram equivalents (MMEs) or prescription counts [@stopka], although these are not directly observable in the present data. Users should choose a unit of analysis that has precedent in their discipline and take appropriate steps to standardize the data (e.g. by population or another stratum) when necessary.
+There are several ways to conceptualize the unit of analysis for opioids from the present data. These include the total number of records, the total number of all opioid pills, the total number of specific opioid pills (i.e. oxycodone versus hydrocodone), or the total amount (in weight) of all or specific opioid pills. Other common units of analysis that may be of interest include morphine milligram equivalents (MMEs) or prescription counts [@stopka]. Users should choose a unit of analysis that has precedent in their discipline and take appropriate steps to standardize the data (e.g. by population or another stratum) when necessary.
 
 ## Conclusion
 
