### 2020-09-08 #2450 @edonnachie
Hi @jeffcsauer ,

Thanks for the detailed response, the updated paper looks good and better explains some of the background. I'm almost through with the review and hope to complete by the end of the week.

@bmcfee : I can't edit the reviewer checklist above; my invitation has expired. Do you need to issue a new invitation or am I missing something?
