### 2020-09-11 #2450 @edonnachie
@bmcfee @jeffcsauer I've now completed my review. The paper, package documentation and functionality are quite impressive. I submitted one minor pull request to bring the documentation up to date  (2014 instead of 2012). Otherwise, I don't find any further issues.
