### 2020-09-11 #2450 @jeffcsauer
Sounds good, I have notified the `arcos` maintainers of your PR and it will be addressed shortly.

Thank you very much for the detailed feedback @edonnachie and @jaeyk! Your review has greatly improved the packages and will surely enhance the experience of future users. We greatly appreciate your time and effort during the JOSS reduced service mode.

If there is anything else we can do, please let us know! 😃
