### 2020-09-14 #2450 @bmcfee
@jeffcsauer the paper looks to be in great shape! :tada:

The last step for you is to make an updated release of the package(s) to reflect any changes made during the review process, and deposit the new version(s) at a DOI-minting repository like Figshare or Zenodo.  When that's done, we can link up the DOIs to the paper and move on to publication.
