### 2020-09-14 #2450 @jeffcsauer
@bmcfee great, sounds good! I have created a Zenodo repository with both `arcos` and `arcospy` [here](https://zenodo.org/record/4029306).
