### 2020-09-15 c813810
Small grammar and typo edits
