### 2020-09-15 #2450 @jeffcsauer
@bmcfee in my final review of the paper I pushed a few small grammar edits, caught a few typos, and added a missing period. Otherwise it is ready on my end! 👍
